requirements.txt
# Agent 系统依赖包
# 使用方式：pip install -r requirements.txt

# ===========================================
# 核心依赖
# ===========================================

# Web 框架
Flask==3.0.0

# API 客户端
openai==1.12.0

# 环境变量管理
python-dotenv==1.0.0

# HTTP 请求
requests==2.31.0

# 网络搜索（可选）
duckduckgo-search==4.1.1

# ===========================================
# 测试依赖
# ===========================================

# 测试框架
pytest==7.4.0

# 测试覆盖率报告
pytest-cov==4.1.0

# ===========================================
# 开发工具（可选）
# ===========================================

# 代码格式化
# black==23.12.0

# 代码检查
# flake8==7.0.0

# 类型检查
# mypy==1.8.0
